+++ packages/aws-client/src/test-utils.ts
@@ -11,7 +11,6 @@ const localStackPorts = {
   stepfunctions: 4566,
   APIGatewayClient: 4566,
   CloudFormation: 4566,
-  cloudwatch: 4566,
   CloudWatchEvents: 4566,
   cloudwatchlogs: 4566,
   DynamoDB: 4566,