+++ packages/aws-client/tests/test-services.js
@@ -90,19 +90,6 @@ test('cloudwatchlogs() service defaults to localstack in test mode', (t) => {
   t.is(cloudwatchlogs.config.endpoint, endpoint);
 });
 
-test('cloudwatch() service defaults to localstack in test mode', (t) => {
-  const cloudwatch = services.cloudwatch();
-  const {
-    credentials,
-    endpoint,
-  } = localStackAwsClientOptions(AWS.CloudWatch);
-  t.deepEqual(
-    cloudwatch.config.credentials,
-    credentials
-  );
-  t.is(cloudwatch.config.endpoint, endpoint);
-});
-
 test('dynamoDb() service defaults to localstack in test mode', async (t) => {
   const dynamodb = services.dynamodb();
   const {