+++ CHANGELOG.md
@@ -43,6 +43,8 @@ terraform values and it will also update the database cluster to use the new con
 
 - **CUMULUS-2888**
   - Update CloudWatch Events code to AWS SDK v3
+- **CUMULUS-2890**
+  - Removed unused CloudWatch AWS SDK client
 - **CUMULUS-2893**
   - Updated Kinesis code to AWS SDK v3